requirements.txt
# Core dependencies
pathlib>=1.0.1
typing-extensions>=4.0.0
python-dotenv>=1.0.0

# LLM API client
openai>=1.0.0

# Data handling
pandas>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
tqdm>=4.65.0  # Progress bars
pyyaml>=6.0.0  # YAML parsing
jsonschema>=4.17.0  # JSON schema validation 